Support PUC-Rio Lua 5.5
Various files within LuaGObject had for loops which reassigned one of
the control variables within the loop's body. Starting in Lua 5.5, this
is considered to be a const reassignment error. This commit fixes all
instances of reassignment of these variables.

Lua 5.5 adds the "global" keyword for lexically scoping a global
variable (by shadowing a local one). To ensure no collisions with
GObject libraries, this commit adds that keyword to LuaGObject's list of
language keywords.

This commit also removes a few non-functional pieces of code.