Finish Adwaita and GTK 4 override
This should nicely finish up the overrides for Adw and GTK 4, while also making sure to test them. This commit removes old GTK 3 tests, which are necessary to test GTK 4. As GTK 3 is obsolete, this is an easy choice to make.

This commit also updates the GTK documentation to note the overridden behaviour, with examples.